Refactor and add path alias
