+++ src/providers/sh/commands/alias/deployment-should-copy-scale.js
@@ -1,0 +1,12 @@
+// @flow
+import type { NpmDeployment, BinaryDeployment } from './types'
+import getScaleForDC from './get-scale-for-dc'
+
+function shouldCopyScalingAttributes(origin: NpmDeployment | BinaryDeployment, dest: NpmDeployment | BinaryDeployment) {
+  return getScaleForDC('bru1', origin).min !== getScaleForDC('bru1', dest).min ||
+    getScaleForDC('bru1', origin).max !== getScaleForDC('bru1', dest).max ||
+    getScaleForDC('sfo1', origin).min !== getScaleForDC('sfo1', dest).min ||
+    getScaleForDC('sfo1', origin).max !== getScaleForDC('sfo1', dest).max
+}
+
+export default shouldCopyScalingAttributes
