+++ src/providers/sh/commands/alias/get-app-name.js
@@ -1,0 +1,26 @@
+// @flow
+import path from 'path'
+import { NowError } from './now-error'
+import { Output } from './types'
+import getConfig from './get-config'
+import readPackage from './read-package'
+
+async function getAppName(output: Output, localConfig: string) {
+  const config = await getConfig(output, localConfig)
+
+  // If the name is in the configuration, return it
+  if (!(config instanceof NowError) && config.name) {
+    return config.name
+  }
+
+  // Otherwise try to get it from the package
+  const pkg = await readPackage()
+  if (!(pkg instanceof NowError) && pkg) {
+    return pkg.name
+  }
+
+  // Finally fallback to directory
+  return path.basename(path.resolve(process.cwd(), localConfig || ''))
+}
+
+export default getAppName
