+++ src/providers/sh/commands/alias/get-deployments-by-appname.js
@@ -1,0 +1,8 @@
+// @flow
+import { Now } from './types'
+
+async function fetchDeploymentsByAppName(now: Now, appName: string) {
+  return now.list(appName, { version: 3 });
+}
+
+export default fetchDeploymentsByAppName
