+++ src/providers/sh/commands/alias/get-domain-status.js
@@ -1,0 +1,13 @@
+// @flow
+import qs from 'querystring'
+import { Now } from './types'
+
+type DomainStatus = {
+  available: boolean,
+}
+
+async function getDomainStatus(now: Now, domain: string): Promise<DomainStatus> {
+  return await now.fetch(`/domains/status?${qs.stringify({ name: domain })}`)
+}
+
+export default getDomainStatus
