+++ src/providers/sh/commands/alias/prompt-bool.js
@@ -1,0 +1,17 @@
+// @flow
+import chalk from 'chalk'
+import { Output } from './types'
+
+async function promptBool(output: Output, message: string) {
+  return new Promise((resolve) => {
+    output.print(`${chalk.gray('>')} ${chalk.bold.red(message)} ${chalk.gray('[y/N] ')}`)
+    process.stdin
+      .on('data', d => {
+        process.stdin.pause()
+        resolve(d.toString().trim().toLowerCase() === 'y')
+      })
+      .resume()
+  })
+}
+
+export default promptBool
