+++ src/providers/sh/commands/alias/setup-domain.js
@@ -1,0 +1,130 @@
+// @flow
+import chalk from 'chalk'
+import psl from 'psl'
+
+// Internal utils
+import domainResolvesToNow from './domain-resolves-to-now'
+import getDomainInfo from './get-domain-info'
+import getDomainNameservers from './get-domain-nameservers'
+import purchaseDomainIfAvailable from './purchase-domain-if-available'
+import setupDNSRecords from './setup-dns-records'
+import verifyDomain from './verify-domain'
+
+// Types and errors
+import { Output, Now } from './types'
+import * as Errors from './errors'
+
+async function setupDomain(output: Output, now: Now, alias: string, contextName: string) {
+  const { subdomain, domain } = psl.parse(alias)
+
+  // In case the domain is avilable, we have to purchase
+  const purchased = await purchaseDomainIfAvailable(output, now, domain, contextName)
+  if ((purchased instanceof Errors.UserAborted) || (purchased instanceof Errors.PaymentSourceNotFound)) {
+    return purchased
+  }
+
+  // Now the domain shouldn't be available and it might or might not belong to the user
+  const info = await getDomainInfo(now, domain, contextName)
+  if (info instanceof Errors.DomainPermissionDenied) {
+    return info
+  }
+
+  if (!info) {
+    output.debug(`Domain is unknown for ZEIT World`)
+    // If we have no info it means that it's an unknown domain. We have to check the
+    // nameservers to register and verify it as an external or non-external domain
+    const nameservers = await getDomainNameservers(now, domain)
+    if (nameservers instanceof Errors.DomainNameserversNotFound) {
+      return nameservers
+    }
+
+    output.log(
+      `Nameservers: ${nameservers && nameservers.length
+        ? nameservers.map(ns => chalk.underline(ns)).join(', ')
+        : chalk.dim('none')}`
+    )
+
+    if (!nameservers.every(ns => ns.endsWith('.zeit.world'))) {
+      // The nameservers are not pointing to zeit world so we have to check if the
+      // DNS are resolving to now. In case they are we can proceed to add and verify
+      if (!await domainResolvesToNow(output, alias, { retries: 5 })) {
+        return new Errors.UnableToResolveDNSExternal(domain, subdomain)
+      }
+
+      // If it doesn't have the nameserver pointing to now we have to create the
+      // domain knowing that it should be verified via a DNS TXT record.
+      const verified = await verifyDomain(now, alias, contextName, { isExternal: true })
+      if (
+        (verified instanceof Errors.DomainNotVerified) ||
+        (verified instanceof Errors.DomainPermissionDenied) ||
+        (verified instanceof Errors.DomainVerificationFailed) ||
+        (verified instanceof Errors.NeedUpgrade)
+      ) {
+        return verified
+      } else {
+        output.success(`Domain ${domain} added!`)
+      }
+    } else {
+      // We have to create the domain knowing that the nameservers are zeit.world
+      output.log(`Detected ${chalk.bold(chalk.underline('zeit.world'))} nameservers! Setting up domain...`)
+      const verified = await verifyDomain(now, alias, contextName, { isExternal: false })
+      if (
+        (verified instanceof Errors.DomainNotVerified) ||
+        (verified instanceof Errors.DomainPermissionDenied) ||
+        (verified instanceof Errors.DomainVerificationFailed) ||
+        (verified instanceof Errors.NeedUpgrade)
+      ) {
+        return verified
+      } else {
+        output.success(`Domain ${domain} added!`)
+      }
+
+      // Since it's pointing to our nameservers we can configure the DNS records
+      const result = await setupDNSRecords(output, now, alias, domain)
+      if (result instanceof Errors.DNSPermissionDenied) {
+        return result
+      }
+
+      // If it doesn't resolve here it means that everything is setup but the
+      // propagation is not done so we invite the user to try later.
+      output.log(`DNS Configured! Verifying propagation…`)
+      if (!await domainResolvesToNow(output, alias, { retries: 10 })) {
+        return new Errors.UnableToResolveDNSInternal(alias)
+      }
+    }
+  } else {
+    // If we have records from the domain we have to try to verify in case it is not
+    // verified and from this point we can be sure about its verification
+    output.debug(`Domain is known for ZEIT World`)
+    if (!info.verified) {
+      const verified = await verifyDomain(now, alias, contextName, { isExternal: info.isExternal })
+      if (
+        (verified instanceof Errors.DomainNotVerified) ||
+        (verified instanceof Errors.DomainPermissionDenied) ||
+        (verified instanceof Errors.DomainVerificationFailed) ||
+        (verified instanceof Errors.NeedUpgrade)
+      ) {
+        return verified
+      }
+    }
+
+    if (!info.isExternal) {
+      // If the domain is an internal domain it means that it's pointing to zeit.world and
+      // we can configure the DNS records in case it is not resolving properly.
+      if (!await domainResolvesToNow(output, alias)) {
+        const result = await setupDNSRecords(output, now, alias, domain)
+        if (result instanceof Errors.DNSPermissionDenied) {
+          return result
+        }
+
+        // Verify that the DNS records are ready
+        output.log(`DNS Configured! Verifying propagation…`)
+        if (!await domainResolvesToNow(output, alias, { retries: 10 })) {
+          return new Errors.UnableToResolveDNSInternal(alias)
+        }
+      }
+    }
+  }
+}
+
+export default setupDomain
